+++ src/org/apache/xerces/xs/XSMultiValueFacet.java
@@ -27,7 +27,7 @@ public interface XSMultiValueFacet extends XSObject {
      * The name of the facet, i.e. <code>FACET_ENUMERATION</code> and 
      * <code>FACET_PATTERN</code> (see <code>XSSimpleTypeDefinition</code>). 
      */
-    public int getFacetKind();
+    public short getFacetKind();
 
     /**
      * Values of this facet.
